Removed quotas set by file and turned them into quotas settable by map. This puts all the quota levels in the scenario yaml which makes it easy to modify for the optimizer. Also worked on a map<-->jtable conversion so that the map can be modified in the GUI.

+++ src/main/java/uk/ac/ox/oxfish/fisher/Fisher.java
@@ -655,7 +655,7 @@ public class Fisher implements Steppable, Startable{
      * @param state the model
      * @return the fish caught and stored (barring overcapacity)
      */
-    public Catch fishHere(GlobalBiology modelBiology, double hoursSpentFishing, FishState state)
+    public Catch fishHere(GlobalBiology modelBiology, int hoursSpentFishing, FishState state)
     {
         Preconditions.checkState(status.getLocation().getAltitude() < 0, "can't fish on land!");
